% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medidor.R
\name{calib}
\alias{calib}
\title{Function to import the calib spreadsheet (.xlsx)}
\usage{
calib(file = "")
}
\arguments{
\item{file}{Full path of the file (.xlsx)}
}
\description{
Function to import the calib spreadsheet (.xlsx)
}
