% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{create_data}
\alias{create_data}
\title{Function to create the spreadsheet (.xlsx)}
\usage{
create_data(segments = 2, path = NULL, path2 = NULL)
}
\description{
Create a 5% or 10% interval spreadsheet
}
\keyword{internal}
