% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medidor.R
\name{import_data}
\alias{import_data}
\title{Function to import the measurement spreadsheet (.xlsx)}
\usage{
import_data(file = "")
}
\arguments{
\item{file}{Full path of the file (.xlsx)}
}
\description{
Function to import the measurement spreadsheet (.xlsx)
}
