% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medidor.R
\name{medidor}
\alias{medidor}
\title{Used to open the MedidoR API
Opens a Shiny application developed to conduct aerial photogrammetry analysis. MedidoR allows you to extract and convert pixel measurements from aerial images collected of marine animals.}
\usage{
medidor()
}
\description{
Used to open the MedidoR API
Opens a Shiny application developed to conduct aerial photogrammetry analysis. MedidoR allows you to extract and convert pixel measurements from aerial images collected of marine animals.
}
