% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medidor.R
\name{medidor_GUI}
\alias{medidor_GUI}
\title{Open MedidoR API}
\usage{
medidor_GUI()
}
\description{
Opens a Shiny application developed to conduct aerial photogrammetry analysis. MedidoR allows you to extract and convert pixel measurements from aerial images collected of marine animals.
}
