% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{show_measurement_modal}
\alias{show_measurement_modal}
\title{show_measurement_modal}
\usage{
show_measurement_modal(message)
}
\description{
Is a helper functions to generate modal dialog
}
\keyword{internal}
