% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{update_measurements}
\alias{update_measurements}
\title{Update the data with calibration model estimates}
\usage{
update_measurements(
  main_path,
  secondary_path,
  model,
  calib_data,
  current_data = NULL
)
}
\description{
Update the data with calibration model estimates
}
\keyword{internal}
