% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{update_measurements}
\alias{update_measurements}
\title{Update the data with calibration model estimates}
\usage{
update_measurements(model, segments)
}
\description{
Update the data with calibration model estimates
}
\keyword{internal}
